requirements.txt
acvl-utils==0.2
addict==2.4.0
albumentations==1.3.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
batchgenerators==0.25
brotlipy==0.7.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.3
colorama==0.4.6
connected-components-3d==3.10.5
contourpy==1.0.7
crcmod==1.7
cryptography @ file:///work/ci_py311_2/cryptography_1679335572456/work
cycler==0.11.0
dicom2nifti==2.4.8
dynamic-network-architectures==0.2
easydict==1.10
elasticdeform==0.5.0
fft-conv-pytorch==1.2.0
filelock==3.14.0
fonttools==4.39.3
fsspec==2023.5.0
future==0.18.3
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
graphviz==0.20.1
hiddenlayer @ git+https://github.com/nanohanno/hiddenlayer.git@321c5d934c6ed3a7e8d330bc4adcf7764bc440c3
huggingface-hub==0.15.1
idna @ file:///work/ci_py311/idna_1676822698822/work
imagecodecs==2023.3.16
imageio==2.28.1
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
jmespath==0.10.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lib==4.0.0
linecache2==1.0.0
lit==16.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.7.1
mdurl==0.1.2
MedPy==0.4.0
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mmcv==2.2.0
mmengine==0.10.5
model-index==0.1.11
monai==1.2.0
mpmath==1.2.1
munch==4.0.0
networkx==3.1
nibabel==5.1.0
# Editable install with no version control (nnunetv2==2.3.1)
-e /data/dpcontact/miniconda3/envs/nnunet/lib/python3.11/site-packages
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.1
Pillow==9.5.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.21.0
pydicom==2.3.1
Pygments==2.18.0
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyparsing==3.0.9
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-gdcm==3.0.22
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
rich==13.4.2
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SimpleITK==2.2.1
six==1.16.0
sympy @ file:///work/ci_py311_2/sympy_1679339311852/work
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traceback2==1.4.0
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1686601443660/work
tzdata==2023.3
unittest2==1.1.0
urllib3 @ file:///croot/urllib3_1686163155763/work
yacs==0.1.8
yapf==0.43.0
